InitialRefactor: Refactor scheduling to its own module to improve readability
